temp/requirements.txt
gitpython
langchain
langchain_community
langchain-google-genai
langchain-huggingface
faiss-cpu
python-dotenv
